后台再修改

+++ bookfun/background/delete_user.php
@@ -57,30 +57,65 @@ form.submit();
                         <td width="47">编辑</td>
                       </tr>
           <?php
-          $link=mysql_connect("localhost","root","zxc7928932")or die("数据库服务器连接错误".mysql_error());//链接数据库
-            mysql_select_db("guestinfo",$link) or die("数据库访问错误".mysql_error());    //选择数据库
-            mysql_query("set names utf-8");                                            //选择字符集标准格式
+          $conn=mysql_connect("localhost","root","zxc7928932") or die("数据库服务器连接错误".mysql_error());
+          mysql_select_db("guestinfo",$conn) or die("数据库访问错误".mysql_error());
+          mysql_query("set names utf-8");
           $keyword=$_POST['txt_keyword'];
-          $sql=mysql_query("select * from register where Username like '%$keyword%' or singing like '%$keyword%'");
-          $row=mysql_fetch_object($sql);
-          if(!$row){
-          /*未查到信息*/
-          echo "<font color='red'>暂无用户信息!</font>";
+          $sql=mysql_query("select * from register where Username like '%$keyword%' or singing like '%$keyword%'");//得到用户信息
+          $message_count=mysql_num_rows($sql);//查询符合条件的记录总条数
+          $row=mysql_fetch_object($sql);//得到对象
+          if($_POST[Submit]!=""){//如果提交搜索 页数初始化为1
+            $page=1;
           }
-          do{
-          ?>
+          /*  $page为当前页，如果$page为空，则初始化为1  */
+          if ($page==""){
+            $page=1;}
+             if (is_numeric($page)){
+            $page_size=4;                     //每页显示4条记录
+            
+            $page_count=ceil($message_count/$page_size);      //根据记录总数除以每页显示的记录数求出所分的页数
+            $offset=($page-1)*$page_size;           //计算下一页从第几条数据开始循环  
+            $sql=mysql_query("select * from register where Username like '%$keyword%' or singing like '%$keyword%' limit $offset,$page_size");
+            $row=mysql_fetch_object($sql);
+            if(!$row){
+              echo "<font color='red'>暂无书籍信息!</font>";
+            }
+            do{
+            ?>
                       <tr bgcolor="#FFFFFF">
-                        <td><?php 
-                        echo $row->Username;
-                        ?></td>
-                        <td><?php echo $row->singing;?></td>
-                        <td align="center"><a href="check_delete_user.php?id=<?php echo $row->uid;?>"><img src="images/delete.gif" width="22" height="22" border="0"></a></td>
+                        <td><?php echo $row->Username ;?></td>
+                        <td><?php echo $row->singing ;?></td>
+                           <td align="center"><a href="check_delete_user.php?id=<?php echo $row->uid;?>"><img src="images/delete.gif" width="22" height="22" border="0"></a></td>
                       </tr>
           <?php
-          }while($row=mysql_fetch_object($sql));
-          mysql_free_result($sql);
-          mysql_close($link);
+            }while($row=mysql_fetch_object($sql));
+          }
           ?>
+                     </table>
+                      <br>
+                      <table width="550" border="0" cellspacing="0" cellpadding="0">
+                        <tr>
+                          <!--  翻页条 -->
+              <td width="37%">&nbsp;&nbsp;页次：<?php echo $page;?>/<?php echo $page_count;?>页&nbsp;记录：<?php echo $message_count;?> 条&nbsp; </td>
+              <td width="63%" align="right">
+              <?php
+              /*  如果当前页不是首页  */
+              if($page!=1){
+              /*  显示“首页”超链接  */
+              echo  "<a href=delete_user.php?page=1>首页</a>&nbsp;";
+              /*  显示“上一页”超链接  */
+              echo "<a href=delete_user.php?page=".($page-1).">上一页</a>&nbsp;";
+              }
+              /*  如果当前页不是尾页  */
+              if($page<$page_count){
+              /*  显示“下一页”超链接  */
+              echo "<a href=delete_user.php?page=".($page+1).">下一页</a>&nbsp;";
+              /*  显示“尾页”超链接  */
+              echo  "<a href=delete_user.php?page=".$page_count.">尾页</a>";
+              }
+              mysql_free_result($sql);
+              mysql_close($conn);
+              ?>
                     </table></td>
                   </tr>
                 </table></td>
